Remove stateless methods into Utils module for now.
To simplify ApiClient’s public API.

Those methods don’t need the client’s state.
In the future those details of HTTP headers handling should be handled internally by ApiClient
