+++ lib/khipu-api-client.rb
@@ -1,4 +1,5 @@
 # Common files
+require 'khipu-api-client/utils'
 require 'khipu-api-client/signature'
 require 'khipu-api-client/api_client'
 require 'khipu-api-client/api_error'
