+++ lib/khipu-api-client/api/banks_api.rb
@@ -16,7 +16,7 @@ module Khipu
       if Configuration.debugging
         Configuration.logger.debug "Calling API: BanksApi#banks_get ..."
       end
-      
+
       # resource path
       path = "/banks".sub('{format}','json')
 
@@ -28,18 +28,18 @@ module Khipu
 
       # HTTP header 'Accept' (if needed)
       _header_accept = ['application/json']
-      _header_accept_result = @api_client.select_header_accept(_header_accept) and header_params['Accept'] = _header_accept_result
+      _header_accept_result = Utils.select_header_accept(_header_accept) and header_params['Accept'] = _header_accept_result
 
       # HTTP header 'Content-Type'
       _header_content_type = ['application/x-www-form-urlencoded']
-      header_params['Content-Type'] = @api_client.select_header_content_type(_header_content_type)
+      header_params['Content-Type'] = Utils.select_header_content_type(_header_content_type)
 
       # form parameters
       form_params = {}
 
       # http body (model)
       post_body = nil
-      
+
 
       auth_names = ['khipu']
       result = @api_client.call_api(:GET, path,
@@ -56,7 +56,4 @@ module Khipu
     end
   end
 end
-
-
-
