+++ lib/khipu-api-client/api/receivers_api.rb
@@ -31,52 +31,52 @@ module Khipu
       if Configuration.debugging
         Configuration.logger.debug "Calling API: ReceiversApi#receivers_post ..."
       end
-      
+
       # verify the required parameter 'admin_first_name' is set
       fail "Missing the required parameter 'admin_first_name' when calling receivers_post" if admin_first_name.nil?
-      
+
       # verify the required parameter 'admin_last_name' is set
       fail "Missing the required parameter 'admin_last_name' when calling receivers_post" if admin_last_name.nil?
-      
+
       # verify the required parameter 'admin_email' is set
       fail "Missing the required parameter 'admin_email' when calling receivers_post" if admin_email.nil?
-      
+
       # verify the required parameter 'country_code' is set
       fail "Missing the required parameter 'country_code' when calling receivers_post" if country_code.nil?
-      
+
       # verify the required parameter 'business_identifier' is set
       fail "Missing the required parameter 'business_identifier' when calling receivers_post" if business_identifier.nil?
-      
+
       # verify the required parameter 'business_category' is set
       fail "Missing the required parameter 'business_category' when calling receivers_post" if business_category.nil?
-      
+
       # verify the required parameter 'business_name' is set
       fail "Missing the required parameter 'business_name' when calling receivers_post" if business_name.nil?
-      
+
       # verify the required parameter 'business_phone' is set
       fail "Missing the required parameter 'business_phone' when calling receivers_post" if business_phone.nil?
-      
+
       # verify the required parameter 'business_address_line_1' is set
       fail "Missing the required parameter 'business_address_line_1' when calling receivers_post" if business_address_line_1.nil?
-      
+
       # verify the required parameter 'business_address_line_2' is set
       fail "Missing the required parameter 'business_address_line_2' when calling receivers_post" if business_address_line_2.nil?
-      
+
       # verify the required parameter 'business_address_line_3' is set
       fail "Missing the required parameter 'business_address_line_3' when calling receivers_post" if business_address_line_3.nil?
-      
+
       # verify the required parameter 'contact_full_name' is set
       fail "Missing the required parameter 'contact_full_name' when calling receivers_post" if contact_full_name.nil?
-      
+
       # verify the required parameter 'contact_job_title' is set
       fail "Missing the required parameter 'contact_job_title' when calling receivers_post" if contact_job_title.nil?
-      
+
       # verify the required parameter 'contact_email' is set
       fail "Missing the required parameter 'contact_email' when calling receivers_post" if contact_email.nil?
-      
+
       # verify the required parameter 'contact_phone' is set
       fail "Missing the required parameter 'contact_phone' when calling receivers_post" if contact_phone.nil?
-      
+
       # resource path
       path = "/receivers".sub('{format}','json')
 
@@ -88,11 +88,11 @@ module Khipu
 
       # HTTP header 'Accept' (if needed)
       _header_accept = ['application/json']
-      _header_accept_result = @api_client.select_header_accept(_header_accept) and header_params['Accept'] = _header_accept_result
+      _header_accept_result = Utils.select_header_accept(_header_accept) and header_params['Accept'] = _header_accept_result
 
       # HTTP header 'Content-Type'
       _header_content_type = ['application/x-www-form-urlencoded']
-      header_params['Content-Type'] = @api_client.select_header_content_type(_header_content_type)
+      header_params['Content-Type'] = Utils.select_header_content_type(_header_content_type)
 
       # form parameters
       form_params = {}
@@ -114,7 +114,7 @@ module Khipu
 
       # http body (model)
       post_body = nil
-      
+
 
       auth_names = ['khipu']
       result = @api_client.call_api(:POST, path,
@@ -131,7 +131,4 @@ module Khipu
     end
   end
 end
-
-
-
